# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Loket API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/event/create": {
            "post": {
                "tags": [
                    "Event"
                ],
                "summary": "Create an event.",
                "description": "Create an event.",
                "operationId": "createEvent",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "name": "eventName",
                        "in": "path",
                        "description": "name of event.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "eventCode",
                        "in": "path",
                        "description": "code of event.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "eventStartDate",
                        "in": "path",
                        "description": "start date of event.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "eventEventDate",
                        "in": "path",
                        "description": "end date of event.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "locationID",
                        "in": "path",
                        "description": "unique location identifier.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "categoryID",
                        "in": "path",
                        "description": "unique category identifier.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "event succesfully created."
                    }
                }
            }
        },
...
    },
...
}